requirements.txt
streamlit
streamlit-option-menu
requests
tavily-python
pandas
altair
python-dotenv
langchain
langchain-groq
langchain-community
faiss-cpu
pypdf
python-docx
openpyxl
huggingface-hub
sentence-transformers
langchain-text-splitters
langchain-huggingface   